refector getting books from backend
