+++ backend/Library/BookService/BookService.csproj
@@ -16,6 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Application\Models\" />
     <Folder Include="Infrastructure\Interfaces\" />
   </ItemGroup>
