+++ backend/Library/BookService/Domain/Models/Book.cs
@@ -2,7 +2,7 @@
 
 public class Book
 {
-    public string Id { get; set; }
+    public Guid Id { get; set; }
     public string Title { get; set; }
     public string Authors { get; set; }
     public string? Description { get; set; }
@@ -10,5 +10,5 @@ public class Book
     public string? Category { get; set; }
     public string Month { get; set; }
     public int Year { get; set; }
-    public float Price { get; set; }
+    public string Price { get; set; }
 }
