+++ backend/Library/BookService/Infrastructure/Repositories/SqlServerBookRepository.cs
@@ -1,7 +1,5 @@
 ﻿using BookService.Domain.Interfaces;
 using BookService.Domain.Models;
-using BookService.Infrastructure.DbModels;
-using Microsoft.EntityFrameworkCore;
 
 namespace BookService.Infrastructure.Repositories;
 
@@ -14,25 +12,8 @@ public class SqlServerBookRepository : IBookRepository
         _databaseContext = databaseContext;
     }
 
-    public async Task<IEnumerable<Book>> GetBooksAsync()
+    public IQueryable<Book> GetBooksAsQuery()
     {
-        var books = await _databaseContext.Books.Take(10).ToListAsync();
-        return books.Select(ConvertDbToDomain).ToArray();
-    }
-
-    private Book ConvertDbToDomain(DbBook book)
-    {
-        return new Book
-        {
-            Id = book.Id.ToString(),
-            Title = book.Title,
-            Authors = book.Authors,
-            Publisher = book.Publisher,
-            Category = book.Category,
-            Description = book.Description,
-            Month = book.Month,
-            Year = book.Year,
-            Price = float.Parse(book.Price)
-        };
+         return  _databaseContext.Books.Take(100);
     }
 }
